requirements.txt
streamlit
pandas
numpy
openpyxyl
python-dotenv
google-generativeai